+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Ability to bulk review multiple commits by selecting a more recent commit on the project page.
+
 ### Fixed
 - Issue with resolving shell's full PATH for built app.
 - LLM was attempting to commit even if no changes were made.